FIX : api directory relocate
