+++ posts/urls.py
@@ -1,12 +1,10 @@
 from django.urls import path
 
-from posts.views import PostView, PostList, TagsView, UserView, UsersView, AuthorView
+from posts.views import PostView, PostList, TagsView, AuthorView
 
 urlpatterns = [
     path('/post/<int:post_id>', PostView.as_view()),
     path('/postlist', PostList.as_view()),
     path('/keyword', TagsView.as_view()),
-    path('/user/<int:user_id>', UserView.as_view()),
-    path('/users/<int:user_id>', UsersView.as_view()),
     path('/authors', AuthorView.as_view()),
 ]
