+++ posts/views.py
@@ -3,7 +3,7 @@ import random
 from django.http  import JsonResponse
 from django.views import View
 
-from users.models import Tag, User
+from users.models import Tag
 from posts.models import Post
 
 class PostView(View):
@@ -48,58 +48,15 @@ class PostList(View):
 
 class TagsView(View):
     def get(self,request):
-        tags    = Tag.objects.filter(usertag__isnull=True).distinct()
+        offset  = int(request.GET.get('offset',0))
+        limit   = int(request.GET.get('limit',24))
+        tags    = Tag.objects.filter(usertag__isnull=True).distinct()[offset:limit]
         results = [{
-            "tag"      : tag.name,
+            "tag"     : tag.name,
             "post_id" : random.choice(list(tag.posts.values_list("id", flat=True)))
         } for tag in tags]
 
         return JsonResponse({"tags" : results}, status = 200)
-
-class UserView(View):
-    def get(self,request,user_id):
-        try:
-            offset  = int(request.GET.get('offset',0))
-            limit   = int(request.GET.get('limit',4))
-            user    = User.objects.all().get(id = user_id)
-            result = [{
-                "profile_image" : user.userimage_set.all().values('url').first(),
-                "name"          : user.author_name,
-                "introduction"  : user.author_intro,
-                "post"          :[{
-                    "id"          : post.id,
-                    "title"       : post.title, 
-                    "sub_title"   : post.sub_title,
-                    "content"     : post.content,
-                    "cover_image" : post.coverimage_set.all().values('url').first()
-                    } for post in user.posts.all()][offset:limit]
-                }]
-             
-            return JsonResponse({"user" : result}, status = 200)
-
-        except KeyError:
-            return JsonResponse({"MESSAGE" : "KEY_ERROR"}, status=400)
-        
-        except User.DoesNotExist:
-            return JsonResponse({"MESSAGE" : "USER_DOES_NOT_EXIST"}, status=404)
-
-class UsersView(View):
-    def get(self,request, user_id):
-        try:
-            user = User.objects.all().get(id = user_id)
-            result = [{
-                "profile_image" : user.userimage_set.all().values('url').first(),
-                "name"          : user.author_name,
-                "introduction"   : user.author_intro,
-            }]
-
-            return JsonResponse({"user" : result}, status = 200)
-
-        except KeyError:
-            return JsonResponse({"MESSAGE" : "KEY_ERROR"}, status=400)
-        
-        except User.DoesNotExist:
-            return JsonResponse({"MESSAGE" : "USER_DOES_NOT_EXIST"}, status=404)
 
 class AuthorView(View):
     def get(self, request):
