+++ users/views.py
@@ -58,3 +58,49 @@ class SignInView(View):
 
         except User.DoesNotExist:
             return JsonResponse({'MESSAGE':'USER_DOES_NOT_EXIST'}, status=404)
+
+
+class UserView(View):
+    def get(self,request,user_id):
+        try:
+            offset  = int(request.GET.get('offset',0))
+            limit   = int(request.GET.get('limit',4))
+            user    = User.objects.all().get(id = user_id)
+            result = [{
+                "profile_image" : user.userimage_set.all().values('url').first(),
+                "name"          : user.author_name,
+                "introduction"  : user.author_intro,
+                "post"          :[{
+                    "id"          : post.id,
+                    "title"       : post.title, 
+                    "sub_title"   : post.sub_title,
+                    "content"     : post.content,
+                    "cover_image" : post.coverimage_set.all().values('url').first()
+                    } for post in user.posts.all()][offset:limit]
+                }]
+             
+            return JsonResponse({"user" : result}, status = 200)
+
+        except KeyError:
+            return JsonResponse({"MESSAGE" : "KEY_ERROR"}, status=400)
+        
+        except User.DoesNotExist:
+            return JsonResponse({"MESSAGE" : "USER_DOES_NOT_EXIST"}, status=404)
+
+class UsersView(View):
+    def get(self,request, user_id):
+        try:
+            user = User.objects.all().get(id = user_id)
+            result = [{
+                "profile_image" : user.userimage_set.all().values('url').first(),
+                "name"          : user.author_name,
+                "introduction"   : user.author_intro,
+            }]
+
+            return JsonResponse({"user" : result}, status = 200)
+
+        except KeyError:
+            return JsonResponse({"MESSAGE" : "KEY_ERROR"}, status=400)
+        
+        except User.DoesNotExist:
+            return JsonResponse({"MESSAGE" : "USER_DOES_NOT_EXIST"}, status=404)
